requirements.txt
# 核心依赖
pybullet>=3.2.5
stable-baselines3>=2.0.0
gymnasium>=1.0.0
torch>=2.0.0

# 数据处理和可视化
numpy>=1.21.0
opencv-python>=4.6.0
matplotlib>=3.5.0
pandas>=1.4.0

# 配置和工具
PyYAML>=6.0
tqdm>=4.64.0
wandb>=0.13.0  # 可选，用于实验跟踪

# 开发依赖
pytest>=7.0.0
black>=22.0.0
isort>=5.10.0
flake8>=4.0.0